pyproject.toml
[project]
name = "contactdb"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django==5.2.*",
    "django-admin-autocomplete-filter>=0.7.1",
    "django-admin-env-notice>=1.0.1",
    "django-auditlog>=3.1.2",
    "django-ckeditor>=6.7.2",
    "django-constance>=4.3.2",
    "django-cors-headers>=4.7.0",
    "django-db-views>=0.1.11",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-group-model>=1.0.1",
    "django-import-export[xlsx]>=4.3.7",
    "django-object-actions>=5.0.0",
    "django-redis>=5.4.0",
    "django-rq>=3.0.1",
    "django-task>=2.0.7",
    "django-two-factor-auth[phonenumbers]>=1.17.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "pycountry>=24.6.1",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "rq-scheduler>=0.14.0",
    "sentry-sdk[django]>=2.28.0",
    "wait-for-it>=2.3.0",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "django-debug-toolbar>=5.2.0",
    "ipdb>=0.13.13",
    "ipython>=9.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.0",
    "pytest-subtests>=0.14.1",
    "ruff>=0.11.9",
]

[tool.ruff.lint]
select = [
    "E4", # pycodestyle
    "E7",
    "E9",
    "F", # pyflakes
    "I", # isort
    "N", # pep8-naming
    "S", # flake8-bandit
    "FBT", # flake8-boolean-trap
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "DJ", # flake8-django
    "LOG", # flake8-logging
    "G", # flake8-logging-format
    "T20", # flake8-print
    "RSE", # flake8-raise
    "RET504", # flake8-return
    "RET505",
    "RET506",
    "RET507",
    "RET508",
    "SIM", # flake8-simplify
    "INT", # flake8-gettext
    "PTH", # flake8-use-pathlib
    "NPY", # numpy rules
    "PT", # flake8-pytest
]
ignore = [
  # Allow boolean default positional arguments, as django has many of those.
  "FBT002",
  # Allow using mark_safe
  "S308",
]

[tool.ruff.lint.per-file-ignores]
"**/conftest.py" = [
  # Allow hardcoded passwords in tests
  "S106",
]
"**/tests/*" = [
  # Allow hardcoded passwords in tests
  "S106",
  # Allow unittest style asserts
  "PT009",
  # Allow unittest style assertRaises
  "PT027",
]
